requirements.txt
Flask==1.0
requests==2.19.0
pyyaml==5.1
